+++ redbot/core/utils/menus.py
@@ -561,3 +561,35 @@ class HybridMenu(dpy_menus.MenuPages, inherit_buttons=False):
     async def stop_pages(self, payload: discord.RawReactionActionEvent) -> None:
         """stops the pagination session."""
         self.stop()
+
+
+class SimpleHybridMenu(HybridMenu, inherit_buttons=True):
+    def __init__(
+        self,
+        source: dpy_menus.PageSource,
+        cog: commands.Cog,
+        clear_reactions_after: bool = True,
+        delete_message_after: bool = True,
+        add_reactions: bool = True,
+        timeout: int = 180,
+        **kwargs: Any,
+    ):
+
+        keyword_to_reaction_mapping = {
+            "last": "\N{BLACK RIGHT-POINTING DOUBLE TRIANGLE WITH VERTICAL BAR}\ufe0f",
+            "first": "\N{BLACK LEFT-POINTING DOUBLE TRIANGLE WITH VERTICAL BAR}\ufe0f",
+            "next": "\N{BLACK RIGHT-POINTING TRIANGLE}\ufe0f",
+            "previous": "\N{BLACK LEFT-POINTING TRIANGLE}\ufe0f",
+            "prev": "\N{BLACK LEFT-POINTING TRIANGLE}\ufe0f",
+            "close": "\N{CROSS MARK}",
+        }
+        super().__init__(
+            source=source,
+            cog=cog,
+            add_reactions=add_reactions,
+            timeout=timeout,
+            clear_reactions_after=clear_reactions_after,
+            delete_message_after=delete_message_after,
+            keyword_to_reaction_mapping=keyword_to_reaction_mapping,
+            **kwargs,
+        )
